New sigOutputStateUpdated behavior
Add ms_on and ms_off to eproc_logic.
Every time a RadioButton is clicked, the signal is sent to the logic and
the update_status comes from there. The only time the update_status
comes from the gui itself is in on_activate.
Add is_eproc_running in eproc_logic.

There are still problems with logic is locked. A solution for this needs
to be found. Add a trigger-like method?